Fix regression tests for Postgres 16+17, move pgstat_custom for clarity
This ensures regression tests pass on Postgres 16 and 17, where the
Plan ID infrastructure is not yet in place, and so currently running
queries don't yet show up in pg_stat_plans. Fixes #4.

In passing, since there was at least one person confused why we were
duplicating the custom stats logic (pgstat_custom.c), move it to a separate
folder that makes it clear this is only for Postgres 16/17 compatibility.
