+++ compat_16_17/expected/activity.out
@@ -1,0 +1,24 @@
+SELECT pg_stat_plans_reset() IS NOT NULL AS t;
+ t 
+---
+ t
+(1 row)
+
+--
+-- check if we see our own plan in activity
+--
+SELECT plan FROM pg_stat_plans_activity WHERE pid = pg_backend_pid();
+ERROR:  Not implemented, use of pg_stat_plans_get_activity requires Postgres 18+
+--
+-- check if we handle showing our plan for named prepared statements correctly
+--
+PREPARE x AS SELECT plan FROM pg_stat_plans_activity WHERE pid = pg_backend_pid();
+EXECUTE x;
+ERROR:  Not implemented, use of pg_stat_plans_get_activity requires Postgres 18+
+DEALLOCATE x;
+SELECT pg_stat_plans_reset() IS NOT NULL AS t;
+ t 
+---
+ t
+(1 row)
+
