+++ compat_16_17/expected/cleanup.out
@@ -1,0 +1,1 @@
+DROP EXTENSION pg_stat_plans;
