+++ pg_stat_plans.c
@@ -39,7 +39,10 @@
 #include "utils/snapmgr.h"
 
 #include "jumblefuncs.h"
-#include "pgstat_custom.h"
+
+#if PG_VERSION_NUM < 180000
+#include "compat_16_17/pgstat_custom.h"
+#endif
 
 #ifdef USE_ZSTD
 #include <zstd.h>
